Frontend of temporarily saving activities
